# contracts/openapi/expense-tracker-api.yaml
openapi: 3.1.0
info:
  title: Expense Tracker API
  version: 1.0.0
  description: Contract for the Expense Tracker backend and Flutter mobile clients.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.example.com
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/categories:
    get:
      tags:
        - Categories
      summary: List categories
      operationId: listCategories
      parameters:
        - $ref: "#/components/parameters/BudgetGroupIdQuery"
        - $ref: "#/components/parameters/TransactionTypeQuery"
      responses:
        "200":
          description: Categories returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryResponse"
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Categories
      summary: Create a category
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryCreateRequest"
      responses:
        "201":
          description: Category created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/v1/reports/summary:
    get:
      tags:
        - Reports
      summary: Get summary report
      operationId: getSummaryReport
      parameters:
        - $ref: "#/components/parameters/BudgetGroupIdQuery"
        - $ref: "#/components/parameters/StartDateQuery"
        - $ref: "#/components/parameters/EndDateQuery"
        - $ref: "#/components/parameters/CategoryIdQuery"
        - $ref: "#/components/parameters/MemberIdQuery"
        - $ref: "#/components/parameters/TransactionTypeQuery"
      responses:
        "200":
          description: Summary report returned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SummaryReportResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/reports/categories:
    get:
      tags:
        - Reports
      summary: Get category breakdown report
      operationId: getCategoryReport
      parameters:
        - $ref: "#/components/parameters/BudgetGroupIdQuery"
        - $ref: "#/components/parameters/StartDateQuery"
        - $ref: "#/components/parameters/EndDateQuery"
        - $ref: "#/components/parameters/CategoryIdQuery"
        - $ref: "#/components/parameters/MemberIdQuery"
        - $ref: "#/components/parameters/TransactionTypeQuery"
      responses:
        "200":
          description: Category report returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryReportItem"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/reports/trends:
    get:
      tags:
        - Reports
      summary: Get expense trend report
      operationId: getExpenseTrendReport
      parameters:
        - $ref: "#/components/parameters/BudgetGroupIdQuery"
        - $ref: "#/components/parameters/StartDateQuery"
        - $ref: "#/components/parameters/EndDateQuery"
        - $ref: "#/components/parameters/CategoryIdQuery"
        - $ref: "#/components/parameters/MemberIdQuery"
        - $ref: "#/components/parameters/TransactionTypeQuery"
      responses:
        "200":
          description: Expense trend report returned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseTrendReportResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/reports/monthly-summary:
    get:
      tags:
        - Reports
      summary: Get monthly summary report
      operationId: getMonthlySummaryReport
      parameters:
        - $ref: "#/components/parameters/BudgetGroupIdQuery"
        - $ref: "#/components/parameters/StartDateQuery"
        - $ref: "#/components/parameters/EndDateQuery"
        - $ref: "#/components/parameters/CategoryIdQuery"
        - $ref: "#/components/parameters/MemberIdQuery"
        - $ref: "#/components/parameters/TransactionTypeQuery"
      responses:
        "200":
          description: Monthly summary report returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlySummaryReportItem"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/reports/members:
    get:
      tags:
        - Reports
      summary: Get member spending report
      operationId: getMemberSpendingReport
      parameters:
        - $ref: "#/components/parameters/BudgetGroupIdQuery"
        - $ref: "#/components/parameters/StartDateQuery"
        - $ref: "#/components/parameters/EndDateQuery"
        - $ref: "#/components/parameters/CategoryIdQuery"
        - $ref: "#/components/parameters/MemberIdQuery"
        - $ref: "#/components/parameters/TransactionTypeQuery"
      responses:
        "200":
          description: Member spending report returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberSpendingReportItem"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/reports/dashboard:
    get:
      tags:
        - Reports
      summary: Get dashboard report
      operationId: getDashboardReport
      parameters:
        - $ref: "#/components/parameters/BudgetGroupIdQuery"
        - $ref: "#/components/parameters/StartDateQuery"
        - $ref: "#/components/parameters/EndDateQuery"
        - $ref: "#/components/parameters/CategoryIdQuery"
        - $ref: "#/components/parameters/MemberIdQuery"
        - $ref: "#/components/parameters/TransactionTypeQuery"
      responses:
        "200":
          description: Dashboard report returned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardReportResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    BudgetGroupIdQuery:
      name: budgetGroupId
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/Uuid"
      description: Filter by budget group identifier.
    CategoryIdQuery:
      name: categoryId
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/Uuid"
      description: Filter by category identifier.
    MemberIdQuery:
      name: memberId
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/Uuid"
      description: Filter by budget member identifier.
    TransactionTypeQuery:
      name: type
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/TransactionType"
      description: Filter by transaction type.
    StartDateQuery:
      name: startDate
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/ApiDate"
      description: Inclusive report or search start date.
    EndDateQuery:
      name: endDate
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/ApiDate"
      description: Inclusive report or search end date.
...
  responses:
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    AuthenticationError:
      description: Authentication failed or token is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    AuthorizationError:
      description: Authenticated user is not allowed to access this resource.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Uuid:
      type: string
      format: uuid
      examples:
        - 0191f4a5-7d0c-7d5b-9f5f-ec31ce70d1a1
    DateTime:
      type: string
      format: date-time
      examples:
        - "2026-06-21T12:00:00Z"
    ApiDate:
      type: string
      format: date
      pattern: "^\\d{4}-\\d{2}-\\d{2}$"
      description: API date value in YYYY-MM-DD format. Flutter clients should collect this with a date picker and display it using the user's locale.
      examples:
        - "2026-06-21"
    MoneyAmount:
      type: number
      format: double
      minimum: 0
      multipleOf: 0.01
      examples:
        - 125.50
    CurrencyCode:
      type: string
      minLength: 3
      maxLength: 3
      pattern: "^[A-Z]{3}$"
      examples:
        - USD
        - TRY
...
    TransactionType:
      type: string
      enum:
        - INCOME
        - EXPENSE
    CategoryCreateRequest:
      type: object
      additionalProperties: false
      required:
        - budgetGroupId
        - name
        - type
      properties:
        budgetGroupId:
          $ref: "#/components/schemas/Uuid"
        name:
          type: string
          minLength: 1
          maxLength: 120
        type:
          $ref: "#/components/schemas/TransactionType"
        color:
          type:
            - string
            - "null"
          pattern: "^#[0-9A-Fa-f]{6}$"
        icon:
          type:
            - string
            - "null"
          maxLength: 80
...
    CategoryResponse:
      type: object
      additionalProperties: false
      required:
        - id
        - budgetGroupId
        - name
        - type
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        budgetGroupId:
          $ref: "#/components/schemas/Uuid"
        name:
          type: string
        type:
          $ref: "#/components/schemas/TransactionType"
        color:
          type:
            - string
            - "null"
        icon:
          type:
            - string
            - "null"
        createdAt:
          $ref: "#/components/schemas/DateTime"
        updatedAt:
          $ref: "#/components/schemas/DateTime"
...
    SummaryReportResponse:
      type: object
      additionalProperties: false
      required:
        - budgetGroupId
        - budgetGroupName
        - startDate
        - endDate
        - totalIncome
        - totalExpense
        - balance
        - currency
        - savingsRate
      properties:
        budgetGroupId:
          $ref: "#/components/schemas/Uuid"
        budgetGroupName:
          type: string
          description: Human-readable budget group name for UI display. Clients must not show budgetGroupId to users.
        startDate:
          $ref: "#/components/schemas/ApiDate"
        endDate:
          $ref: "#/components/schemas/ApiDate"
        totalIncome:
          $ref: "#/components/schemas/MoneyAmount"
        totalExpense:
          $ref: "#/components/schemas/MoneyAmount"
        balance:
          type: number
          format: double
          multipleOf: 0.01
        currency:
          $ref: "#/components/schemas/CurrencyCode"
        savingsRate:
          type: number
          format: double
          minimum: 0
          maximum: 100
          description: Savings rate as a percentage of income after expenses.
    CategoryReportItem:
      type: object
      additionalProperties: false
      required:
        - categoryId
        - categoryName
        - totalAmount
        - percentage
        - currency
      properties:
        categoryId:
          $ref: "#/components/schemas/Uuid"
        categoryName:
          type: string
        transactionType:
          $ref: "#/components/schemas/TransactionType"
        totalAmount:
          $ref: "#/components/schemas/MoneyAmount"
        percentage:
          type: number
          format: double
          minimum: 0
          maximum: 100
          description: Percentage share within the filtered report total.
        currency:
          $ref: "#/components/schemas/CurrencyCode"
    ExpenseTrendReportResponse:
      type: array
      items:
        $ref: "#/components/schemas/ExpenseTrendReportItem"
    ExpenseTrendReportItem:
      type: object
      additionalProperties: false
      required:
        - date
        - income
        - expense
        - balance
        - currency
      properties:
        date:
          $ref: "#/components/schemas/ApiDate"
        income:
          $ref: "#/components/schemas/MoneyAmount"
        expense:
          $ref: "#/components/schemas/MoneyAmount"
        balance:
          type: number
          format: double
          multipleOf: 0.01
        currency:
          $ref: "#/components/schemas/CurrencyCode"
    MonthlySummaryReportItem:
      type: object
      additionalProperties: false
      required:
        - month
        - income
        - expense
        - balance
        - currency
      properties:
        month:
          type: string
          pattern: "^\\d{4}-\\d{2}$"
          description: Month in YYYY-MM format.
        income:
          $ref: "#/components/schemas/MoneyAmount"
        expense:
          $ref: "#/components/schemas/MoneyAmount"
        balance:
          type: number
          format: double
          multipleOf: 0.01
        currency:
          $ref: "#/components/schemas/CurrencyCode"
    MemberSpendingReportItem:
      type: object
      additionalProperties: false
      required:
        - memberId
        - memberName
        - amount
        - percentage
        - currency
      properties:
        memberId:
          $ref: "#/components/schemas/Uuid"
        memberName:
          type: string
          description: Human-readable member name for UI display. Clients must not show memberId to users.
        amount:
          $ref: "#/components/schemas/MoneyAmount"
        percentage:
          type: number
          format: double
          minimum: 0
          maximum: 100
        currency:
          $ref: "#/components/schemas/CurrencyCode"
    DashboardReportResponse:
      type: object
      additionalProperties: false
      required:
        - budgetGroupId
        - budgetGroupName
        - startDate
        - endDate
        - currency
        - summary
      properties:
        budgetGroupId:
          $ref: "#/components/schemas/Uuid"
        budgetGroupName:
          type: string
          description: Human-readable budget group name for UI display. Clients must not show budgetGroupId to users.
        startDate:
          $ref: "#/components/schemas/ApiDate"
        endDate:
          $ref: "#/components/schemas/ApiDate"
        currency:
          $ref: "#/components/schemas/CurrencyCode"
        summary:
          $ref: "#/components/schemas/DashboardSummaryWidgets"
        categoryBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/CategoryReportItem"
        trend:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseTrendReportItem"
        monthlySummary:
          type: array
          items:
            $ref: "#/components/schemas/MonthlySummaryReportItem"
        memberSpending:
          type: array
          items:
            $ref: "#/components/schemas/MemberSpendingReportItem"
    DashboardSummaryWidgets:
      type: object
      additionalProperties: false
      required:
        - totalIncome
        - totalExpense
        - balance
        - savingsRate
        - transactionCount
      properties:
        currency:
          $ref: "#/components/schemas/CurrencyCode"
        totalIncome:
          $ref: "#/components/schemas/MoneyAmount"
        totalExpense:
          $ref: "#/components/schemas/MoneyAmount"
        balance:
          type: number
          format: double
          multipleOf: 0.01
        savingsRate:
          type: number
          format: double
          minimum: 0
          maximum: 100
        transactionCount:
          type: integer
          format: int64
          minimum: 0
        topCategory:
          $ref: "#/components/schemas/CategoryReportItem"
        topMember:
          $ref: "#/components/schemas/MemberSpendingReportItem"
...
    FieldError:
      type: object
      additionalProperties: false
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - timestamp
        - status
        - errorCode
        - message
        - details
      properties:
        timestamp:
          $ref: "#/components/schemas/DateTime"
        status:
          type: integer
          format: int32
        errorCode:
          type: string
          examples:
            - VALIDATION_ERROR
            - INVALID_CREDENTIALS
            - ACCESS_DENIED
            - RESOURCE_NOT_FOUND
            - INTERNAL_SERVER_ERROR
        message:
          type: string
        details:
          type: array
          description: Field-level or contextual details. Use an empty array when no details are available.
          items:
            $ref: "#/components/schemas/FieldError"